move A to fix broken testisGeneric
