+++ src/test/java/spoon/test/generics/testclasses4/A.java
@@ -1,4 +1,4 @@
-package spoon.test.generics.testclasses.overriding;
+package spoon.test.generics.testclasses4;
 
 public class A {
     <T extends Enum<T>> void m6(T t) {
